% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{bigene_getValues}
\alias{bigene_getValues}
\title{bi_getValues utility function.}
\usage{
bigene_getValues(object, features, limita, limitb)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Compute N PC dims}
}
\description{
bi_getValues utility function.
}
