% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{celltypeVlnPlot}
\alias{celltypeVlnPlot}
\title{Function to plot multiple genes in a single Violin plot}
\usage{
celltypeVlnPlot(
  object,
  features,
  group.by = "var_cluster",
  cols,
  orientation = "vertical"
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{genes to plot}

\item{group.by}{variable to group the cells by}

\item{cols}{Colors}

\item{orientation}{orientation to order the plots horizontally (single row) or vertically (single column)}
}
\description{
Function to plot multiple genes in a single Violin plot
}
