% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDR.R
\name{ClusterDR}
\alias{ClusterDR}
\title{Run all Dimension Reduction methods and find marker genes between clusters}
\usage{
ClusterDR(object, n.neighbors = 30, dims, DM = F, UMAP = T,
  TSNE = T, findallmarkers = T, resolution = 0.5, n.components = 2,
  min.dist = 0.3)
}
\arguments{
\item{object}{Seurat object}

\item{n.neighbors}{n number of neighbors for umap}

\item{DM}{Run diffusion map}

\item{UMAP}{Run UMAP}

\item{TSNE}{Run TSNE}

\item{findallmarkers}{T/F run findallmarkers}

\item{resolution}{Resolution param for FindCluster}

\item{n.components}{How many components to compute for Dim reductions (UMAP,DM,etc)}

\item{min.dist}{minimum distance parameter for RunUMAP function. Controls how tightly to compress the umap points together}
}
\description{
Run all Dimension Reduction methods and find marker genes between clusters
}
