% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{CurvePlot}
\alias{CurvePlot}
\title{CurvePlot Plot Slingshot Curves}
\usage{
CurvePlot(
  object,
  sds = NULL,
  group.by = NULL,
  reduction = "umap",
  lineage = "all",
  dims = 1:2,
  cols = NULL,
  label = T
)
}
\arguments{
\item{object}{Seurat object}

\item{sds}{Slingshot Data object}

\item{group.by}{variable to group by}

\item{reduction}{Which dimensionality reduction to use, default UMAP}

\item{lineage}{Linage to plot or all to plot all lineages}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{cols}{Color palette}

\item{label}{Label plots}
}
\description{
CurvePlot Plot Slingshot Curves
}
