% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{DimPCAPlot}
\alias{DimPCAPlot}
\title{Plot Pairwise PCA plots.}
\usage{
DimPCAPlot(object, dims = 1:10, feature)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{Compute N PC dims}
}
\description{
Plot Pairwise PCA plots.
}
