% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlotTopGenes.R
\name{DotPlotTopGenes}
\alias{DotPlotTopGenes}
\title{Plot dotplot of top n genes}
\usage{
DotPlotTopGenes(object, nfeatures = 3)
}
\arguments{
\item{object}{Seurat object}

\item{nfeatures}{number of genes to plot. Number has to be between 1 and 10.}
}
\description{
Plot dotplot of top n genes
}
