% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{FeaturePlot2}
\alias{FeaturePlot2}
\title{Visualize 'features' on a dimensional reduction plot}
\usage{
FeaturePlot2(object, features, dims = c(1, 2), cells = NULL,
  cols = c("lightgrey", "blue"), pt.size = NULL, order = TRUE,
  min.cutoff = NA, max.cutoff = NA, reduction = NULL,
  split.by = NULL, shape.by = NULL, blend = FALSE,
  blend.threshold = 0.5, label = FALSE, label.size = 4,
  ncol = NULL, combine = TRUE, coord.fixed = FALSE, by.col = TRUE)
}
\arguments{
\item{features}{Vector of features to plot.}

\item{cols}{The two colors to form the gradient over. Provide as string vector with
the first color corresponding to low values, the second to high. Also accepts a Brewer
color scale or vector of colors. Note: this will bin the data into number of colors provided.}

\item{order}{Boolean determining whether to plot cells in order of expression. Can be useful if
cells expressing given feature are getting buried.}

\item{min.cutoff, max.cutoff}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}

\item{split.by}{A factor in object metadata to split the feature plot by, pass 'ident'
to split by cell identity'; similar to the old \code{FeatureHeatmap}}

\item{blend}{Scale and blend expression values to visualize coexpression of two features}

\item{blend.threshold}{The color cutoff from weak signal to strong signal; ranges from 0 to 1.}

\item{ncol}{Number of columns to combine multiple feature plots to, ignored if \code{split.by} is not \code{NULL}}

\item{combine}{Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features}

\item{coord.fixed}{Plot cartesian coordinates with fixed aspect ratio}

\item{by.col}{If splitting by a factor, plot the splits per column with the features as rows.}
}
\value{
A ggplot object
}
\description{
Colors single cells on a dimensional reduction plot according to a 'feature'
(i.e. gene expression, PC scores, number of genes detected, etc.)
}
\examples{
FeaturePlot(object = pbmc_small, features = 'PC_1')
}
