% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterMarkers.R
\name{getClusterMarkers}
\alias{getClusterMarkers}
\title{Get marker genes for the given cluster}
\usage{
getClusterMarkers(object, cluster = 0)
}
\arguments{
\item{object}{Seurat object}

\item{cluster}{name of cluster you want markers for}
}
\description{
Get marker genes for the given cluster
}
