% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{getSDS}
\alias{getSDS}
\title{getSDS Plot Slingshot Curves}
\usage{
getSDS(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns an object of class SlingshotDataSet
}
\description{
getSDS Plot Slingshot Curves
}
