% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTopGenes.R
\name{HeatMapTopGenes}
\alias{HeatMapTopGenes}
\title{Plot heatmap of top n genes}
\usage{
HeatMapTopGenes(object, nfeatures = 10)
}
\arguments{
\item{object}{Seurat object}

\item{nfeatures}{number of features to plot. Default is 10}
}
\description{
Plot heatmap of top n genes
}
