% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligrec.R
\name{ligrec}
\alias{ligrec}
\title{Get ligand receptor pairs between all clusters}
\usage{
ligrec(object, grp.var = "ident", org, perc = 30)
}
\arguments{
\item{object}{Seurat object}

\item{grp.var}{grouping variable}

\item{org}{organism}

\item{perc}{minimum percentage of cells that should express the ligand or receptor}
}
\description{
Get ligand receptor pairs between all clusters
}
