% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{lineageDimPlot}
\alias{lineageDimPlot}
\title{lineageDimPlot Plot Slingshot Curves}
\usage{
lineageDimPlot(
  object,
  reduction = "umap",
  group.by = NULL,
  lineage = "all",
  dims = 1:2,
  cols = NULL,
  label = T
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduction used}

\item{group.by}{variable to group by}

\item{lineage}{Linage to plot or all to plot all lineages}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{cols}{Color palette}

\item{label}{Label plots}
}
\description{
lineageDimPlot Plot Slingshot Curves
}
