% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{LineageFeaturePlot}
\alias{LineageFeaturePlot}
\title{LineageFeaturePlot Plot Slingshot Curves}
\usage{
LineageFeaturePlot(
  object,
  lineage = "lineage1",
  reduction = "umap",
  dims = 1:2,
  cols = "RdYlBu",
  group.by = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Which dimensionality reduction to use, default UMAP}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{cols}{Color palette}

\item{group.by}{Cluster variable used in slingshot}
}
\description{
LineageFeaturePlot Plot Slingshot Curves
}
