% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make3dPlot.R
\name{make3dPlot}
\alias{make3dPlot}
\title{Create a 3D plot}
\usage{
make3dPlot(object, groupby, reduction = "dm", colors = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{groupby}{grouping variable to color by. Has to be a column name present in metadata of the seurat object}

\item{reduction}{Reduction method. Default is 'dm'}

\item{colors}{color palette}
}
\description{
Create a 3D plot
}
