% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PCAPwPlot}
\alias{PCAPwPlot}
\title{Plot Pairwise PCA plots.}
\usage{
PCAPwPlot(object, dims = 1:50)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{Compute N PC dims}
}
\description{
Plot Pairwise PCA plots.
}
