% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDR.R
\name{PCATools}
\alias{PCATools}
\title{Run PCA and Jackstaw and plots.}
\usage{
PCATools(object, npcs = 50, jackstraw = T, plotdir = "./")
}
\arguments{
\item{object}{Seurat object}

\item{npcs}{Compute N PC dims}

\item{jackstraw}{Run jackstraw}

\item{plotdir}{Where to export plots.}
}
\description{
Run PCA and Jackstaw and plots.
}
