% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{plotCurveDGEgenes}
\alias{plotCurveDGEgenes}
\title{Create feature plots of genes}
\usage{
plotCurveDGEgenes(object = NULL, curve = NULL, n = 25,
  reduction.use = "dm")
}
\arguments{
\item{object}{Seurat object}

\item{curve}{curve to be plotted}

\item{n}{number of genes to plot. Default is 25}

\item{reduction.use}{Reduction method. Default is 'dm'}
}
\description{
Create feature plots of genes
}
