% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{plotLineageHeatMap}
\alias{plotLineageHeatMap}
\title{Plot heatmap of the pseudotime data}
\usage{
plotLineageHeatMap(
  object,
  features,
  lineage = "lineage1",
  col,
  group.by = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Vector of genes to be plotted}

\item{lineage}{THe linage to be plotted such as lineage1}

\item{col}{Color palette, this vector needs to have the names be the cell types or cluster names}

\item{group.by}{Cluster variable used in slingshot}

\item{sdsname}{Name of Slingshot object stored in Seurat Object}
}
\description{
Plot heatmap of the pseudotime data
}
