% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{plotLineangeGene}
\alias{plotLineangeGene}
\title{Plot heatmap of the pseudotime data}
\usage{
plotLineangeGene(
  object,
  features,
  lineage = "lineage1",
  col,
  lo.col = "red",
  hi.col = "green",
  group.by = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Vector of genes to be plotted}

\item{lineage}{THe linage to be plotted such as lineage1}

\item{lo.col}{Low value color}

\item{hi.col}{high value color}

\item{group.by}{Cluster variable used in slingshot}
}
\description{
Plot heatmap of the pseudotime data
}
