% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRnetwork.R
\name{plotLRnetwork}
\alias{plotLRnetwork}
\title{Create ligand receptor network plot}
\usage{
plotLRnetwork(results, filtermin = 0, filtermax = 0)
}
\arguments{
\item{results}{dataframe containing the fields- interacting_pair,Receptor,Ligand,Receptor_cluster,Ligand_cluster}

\item{filtermin}{minimum number of interactions between the nodes}

\item{filtermax}{maximum number of interactions between the nodes}
}
\value{
Returns network plot with ligand-receptor pair data
}
\description{
Create ligand receptor network plot
}
