% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{plotPseudoTime}
\alias{plotPseudoTime}
\title{run Pseudotime}
\usage{
plotPseudoTime(
  object,
  sds = NULL,
  group.by = NULL,
  reduction = "dm",
  dims = 1:2
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{variable to group by}

\item{reduction}{Which dimensionality reduction to use}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}
}
\description{
run Pseudotime
}
