% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processExper.R
\name{processExper}
\alias{processExper}
\title{Preprocess single cell RNA-seq data with seurat and return seurat object. This function only does normalization, finding variable genes and scaling}
\usage{
processExper(object, ccscale = F, return_var_genes = F, org)
}
\arguments{
\item{object}{Seurat object}

\item{ccscale}{Specify whether to scale for cell cycle genes or not}

\item{return_var_genes}{Define if scale.data matrix should contain only the variable genes. T/F}

\item{organism}{Organism (mouse/human)}
}
\value{
Seurat object
}
\description{
Preprocess single cell RNA-seq data with seurat and return seurat object. This function only does normalization, finding variable genes and scaling
}
\examples{
scrna = processExper(object=scrna,ccscale=F,sc.transform =F)
}
