% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSeuratObj.R
\name{projectname Project name}
\alias{projectname Project name}
\alias{CreateSeuratObj}
\title{Create Seurat object from 10x output}
\usage{
CreateSeuratObj(files, name)
}
\arguments{
\item{files}{input files}
}
\value{
Seurat object
}
\description{
Create Seurat object from 10x output
}
\examples{
scrna = CreateSeuratObj(dir=prjdir,name='test_prj',files=c("10x/Gene/filtered"))
}
