% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiffusion.R
\name{RunDiffusion}
\alias{RunDiffusion}
\title{Run Diffusion map}
\usage{
RunDiffusion(
  object,
  dims = 1:5,
  reduction = "pca",
  features = NULL,
  assay = "RNA",
  max.dim = 3L,
  q.use = 0.01,
  reduction.name = "dm",
  reduction.key = "DM_",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{number of dimensions}

\item{reduction}{reductionm method. Defaults to pca}

\item{features}{vector of gene names}

\item{assay}{Assay name. Defaults to 'RNA'}

\item{max.dim}{maximum dimensions}

\item{q.use}{q. value cutoff}

\item{reduction.name}{Dimension Reduction method}

\item{reduction.key}{Dimension Reduction key}
}
\description{
Run Diffusion map
}
