% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDoubletfinder.R
\name{RunDoubletfinder}
\alias{RunDoubletfinder}
\title{Function to run doublet finder}
\usage{
RunDoubletfinder(
  object,
  dims,
  group = "var_cluster",
  doublet.formrate = 0.75,
  sct = T
)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{Number of dimensions to use}

\item{group}{variable to group the cells by}

\item{doublet.formrate}{Doublet formation rate (Refer to Cellranger documentation. Depends on number of cells fed in)}

\item{sct}{logical (T/F)}
}
\description{
Function to run doublet finder
}
