% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{runPseudoTimeDGE}
\alias{runPseudoTimeDGE}
\title{run Psuedotime Diff Expression}
\usage{
runPseudoTimeDGE(object)
}
\arguments{
\item{object}{Seurat object}
}
\description{
run Psuedotime Diff Expression
}
