% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQC.R
\name{RunQC}
\alias{RunQC}
\title{Perform QC on single cell RNA-seq data with seurat, store all stats and return seurat object}
\usage{
RunQC(
  dir,
  name,
  org = "mouse",
  files,
  filter = T,
  LowerFeatureCutoff = 200,
  UpperFeatureCutoff = "MAD",
  UpperMitoCutoff = 5,
  doubletdetection = F
)
}
\arguments{
\item{dir}{Directory containing the cellranger files for input}

\item{org}{Organism}

\item{filter}{Specify whether to filter data}

\item{LowerFeatureCutoff}{Lower value for Feature cutoff}

\item{UpperMitoCutoff}{Upper Mito Pecent cuoff default is 0.05}

\item{doubletdetection}{Logical to specify if you want to run scds doublet detection}

\item{ccscale}{Specify whether to scale for cell cycle genes or not}

\item{UpperfeatureCutoff}{Upper bound, can be an integer or if set to MAD will be 3 median + 3*MAD}

\item{sc.transform}{Define whether or not to apply scTransform function in seurat. T/F}
}
\value{
Seurat object
}
\description{
Perform QC on single cell RNA-seq data with seurat, store all stats and return seurat object
}
\examples{
scrna = processExper(dir=prjdir,'test_prj',org='mouse',files=c("filtered_gene_bc_matrices/mm10"),ccscale=F,filter = T,LowerFeatureCutoff=200,UpperFeatureCutoff="MAD",UpperMitoCutoff=0.05)
}
