% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{runSlingshot}
\alias{runSlingshot}
\title{run Slingshot}
\usage{
runSlingshot(
  object,
  reduction = "dm",
  group.by = NULL,
  start.clus = NULL,
  end.clus = NULL,
  approx_points = FALSE,
  allow.breaks = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduction method. Default is 'dm'}

\item{start.clus}{start cluster}

\item{end.clus}{end cluster}

\item{group, by}{groups}
}
\description{
run Slingshot
}
