% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{sampleBarGraph}
\alias{sampleBarGraph}
\title{Plot a stacked bar graph of per Sample contribtion}
\usage{
sampleBarGraph(object, group.by = NULL, col = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Variable for x axis typicall cluster or celltype}

\item{cols}{Color palette for grouping variable}
}
\description{
Plot a stacked bar graph of per Sample contribtion
}
