% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{setSDS}
\alias{setSDS}
\title{setSDS Plot Slingshot Curves}
\usage{
setSDS(object, sds)
}
\arguments{
\item{object}{Seurat object}

\item{sds}{A SlingshotDataSet object}
}
\value{
Returns an object of class Seurat with SDS set into the misc slot.
}
\description{
setSDS Plot Slingshot Curves
}
