% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{SplitMetaPlot}
\alias{SplitMetaPlot}
\title{Function to plot}
\usage{
SplitMetaPlot(object, reduction = "umap", feature = NULL, ncol = 4, col = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduction used for the plot, umap default}

\item{feature}{Feature to split on, ie sample}

\item{ncol}{How many columns to use for the plot}

\item{cols}{Color palette}
}
\value{
patchwork object
}
\description{
Function to plot
}
