% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{BiGenePlot}
\alias{BiGenePlot}
\title{bi_getValues utility function.}
\usage{
BiGenePlot(object, features, x = 1, y = 2, limita = c(1, 100),
  limitb = c(1, 100), pt.size = 0.1, title = NULL,
  reduction = "umap")
}
\arguments{
\item{object}{Seurat object}

\item{features}{2 genes}

\item{pt.size}{Adjust point size}

\item{title}{Plot Tttle}

\item{reduction}{Default is umap}
}
\description{
bi_getValues utility function.
}
