% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDR.R
\name{ClusterDR}
\alias{ClusterDR}
\title{Run all Dimension Reduction methods and find marker genes between clusters}
\usage{
ClusterDR(object, npcs = 50, maxdim = "auto", k = 30)
}
\arguments{
\item{object}{Seurat object}

\item{npcs}{number of PC's to run}

\item{maxdim}{maximum number of dimensions. Defaults to npcs specified}

\item{k}{n number of neighbors for umap}
}
\description{
Run all Dimension Reduction methods and find marker genes between clusters
}
