% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDR.R
\name{ClusterDR}
\alias{ClusterDR}
\title{Run all Dimension Reduction methods and find marker genes between clusters}
\usage{
ClusterDR(object, k = 30, dims = dim, DM = F, UMAP = T, TSNE = T,
  resolution = 0.5)
}
\arguments{
\item{object}{Seurat object}

\item{k}{n number of neighbors for umap}

\item{DM}{Run diffusion map}

\item{UMAP}{Run UMAP}

\item{TSNE}{Run TSNE}

\item{resolution}{Resolution param for FindCluster}
}
\description{
Run all Dimension Reduction methods and find marker genes between clusters
}
