% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{plotCurveHeatmaps}
\alias{plotCurveHeatmaps}
\title{Plot heatmap of the pseudotime data}
\usage{
plotCurveHeatmaps(object = NULL, curve = NULL,
  annCol = "var_cluster", filename = "heatmap.png", n = 25)
}
\arguments{
\item{object}{Seurat object}

\item{curve}{curve to be plotted}

\item{filename}{output filename}

\item{n}{number of genes to plot. Default is 25}
}
\description{
Plot heatmap of the pseudotime data
}
