% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlingShot.R
\name{plotLineageHeatMap}
\alias{plotLineageHeatMap}
\title{Plot heatmap of the pseudotime data}
\usage{
plotLineageHeatMap(object, sdsname, features, lineage = "lineage1", col)
}
\arguments{
\item{object}{Seurat object}

\item{sdsname}{Name of Slingshot object stored in Seurat Object}

\item{features}{Vector of genes to be plotted}

\item{lineage}{THe linage to be plotted such as lineage1}

\item{col}{Color palette, this vector needs to have the names be the cell types or cluster names}
}
\description{
Plot heatmap of the pseudotime data
}
