% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudoTime.R
\name{plotPseudoTime}
\alias{plotPseudoTime}
\title{run Pseudotime}
\usage{
plotPseudoTime(object, group.by = NULL, reduction = "dm", dims = 1:2)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{variable to group by}

\item{reduction}{Reduction method. Default is 'dm'}

\item{dims}{number of dimensions}
}
\description{
run Pseudotime
}
