% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processExper.R
\name{processExper}
\alias{processExper}
\title{Preprocess single cell RNA-seq data with seurat and return seurat object}
\usage{
processExper(dir, name, org = "mouse", files, ccscale = F,
  filter = T)
}
\arguments{
\item{dir}{Directory containing the cellranger files for input}

\item{name}{Project name}

\item{org}{Organism}

\item{files}{path to the dir having the barcodes.tsv, genes.tsv and matrix.mtx files}

\item{ccscale}{Specify whether to scale for cell cycle genes or not}

\item{filter}{Specify whether to filter data}
}
\value{
Seurat object
}
\description{
Preprocess single cell RNA-seq data with seurat and return seurat object
}
\examples{
scrna = processExper(dir=prjdir,'test_prj',org='mouse',files=c("filtered_gene_bc_matrices/mm10"),ccscale=F,filter = T)
}
