% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQC.R
\name{RunQC}
\alias{RunQC}
\title{Perform QC on single cell RNA-seq data with seurat, store all stats and return seurat object}
\usage{
RunQC(
  object,
  org = "mouse",
  filter = T,
  LowerFeatureCutoff = 200,
  UpperFeatureCutoff = "MAD",
  UpperMitoCutoff = 5,
  doubletdetection = T,
  dir
)
}
\arguments{
\item{object}{Seurat object}

\item{org}{Organism (mouse/human)}

\item{filter}{Specify whether to filter data}

\item{LowerFeatureCutoff}{Lower value for Feature cutoff}

\item{UpperMitoCutoff}{Upper Mito Pecent cuoff default is 0.05}

\item{doubletdetection}{Logical to specify if you want to run scds doublet detection}

\item{dir}{Output directory for QC files}

\item{UpperfeatureCutoff}{Upper bound, can be an integer or if set to MAD will be 3 median + 3*MAD}
}
\value{
Seurat object
}
\description{
Perform QC on single cell RNA-seq data with seurat, store all stats and return seurat object
}
\examples{
scrna = RunQC(object=scrna,org="mouse",filter = T,LowerFeatureCutoff=200,UpperFeatureCutoff="MAD",UpperMitoCutoff=0.05,doubletdetection = T)
}
