% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSlingshot.R
\name{runSlingshot}
\alias{runSlingshot}
\title{run Slingshot}
\usage{
runSlingshot(object, reduction = "dm", groups = NULL,
  start.clus = NULL, end.clus = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduction method. Default is 'dm'}
}
\description{
run Slingshot
}
