% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{sampleBarGraph}
\alias{sampleBarGraph}
\title{Plot a stacked bar graph of}
\usage{
sampleBarGraph(object, group.by = NULL, color.by = NULL, col = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Variable for x axis typical cluster or celltype}

\item{color.by}{Variable to split each group by such as library}

\item{cols}{Color palette for grouping variable}
}
\description{
Plot a stacked bar graph of
}
